% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{ac.data}
\alias{ac.data}
\title{Air Conditioning data from EIA survey}
\format{An object of class \code{data.frame} with 3108 rows and 6 columns.}
\source{
\url{https://www.eia.gov/consumption/residential/data/2015/}, 
\url{https://www.arcgis.com/home/item.html?id=8e5c3c6e1fa94e379553e199dcc4e777#overview}
}
\usage{
ac.data
}
\description{
Air conditioning data were taken from the Residential Energy Consumption Survey (RECS) from the U.S. Energy Information Administration
Data averaged from 2009 and 2015. Data were also available for 2001 and 2005, but used a different climate region system
Data were joined spatially based on a Climate region shapefile (see below).
}
\details{
See wnv_data_dev.R for details on data processing. Averages were computed in MS Excel.

The data layer used for spatial information was from the Building America website:
https://www.arcgis.com/home/item.html?id=8e5c3c6e1fa94e379553e199dcc4e777#overview
Details on the climate zone methods are here: https://www.energy.gov/sites/prod/files/2015/10/f27/ba_climate_region_guide_7.3.pdf
Climate Zones also appear to be available here: https://www.eia.gov/maps/layer_info-m.php
}
\keyword{datasets}
