% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{census.data}
\alias{census.data}
\title{Census Data by year}
\format{An object of class \code{data.frame} with 62160 rows and 7 columns.}
\source{
#**# FILL IN
}
\usage{
census.data
}
\description{
Census data #**# FILL IN DOCUMENTATION
}
\keyword{datasets}
