% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{fips.lookup}
\alias{fips.lookup}
\title{FIPS Lookup Table}
\format{An object of class \code{data.frame} with 3108 rows and 4 columns.}
\source{
U.S. Census Bureau #**# WHERE
}
\usage{
fips.lookup
}
\description{
FIPS codes and CDC locations to facilitate merging data sets by either FIPS or by location
See wnv_data_dev.R for minor modifications to the Census data set
}
\keyword{datasets}
