% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{mosquito.ranges}
\alias{mosquito.ranges}
\title{Mosquito Ranges}
\format{An object of class \code{data.frame} with 2821 rows and 9 columns.}
\source{
Bernard, K. and Kramer L. 2001 West Nile virus activity in the United States, 2001. Viral Immunology 14: 319-338.
Darsie, R.F., Jr., and R.A. Ward. 1981. Identification and geographical distribution of the mosquitoes of North America,
north of Mexico. Supplement to Mosquito Systematics 1:1-313.
Darsie, R.F., Jr., and R.A. Ward. 1989. Review of new Nearctic mosquito distributional records north of Mexico,
with notes on additions and taxonomic changes of the fauna, 1982-89. J. Am. Mosq. Control Assoc. 5:552-557
}
\usage{
mosquito.ranges
}
\description{
Merge by fips column, note that location column is not present
Mosquito Range Maps were digitized from Kramer and Bernard 2001 using ArcGIS.
County level data used counties form the U.S. Census (tl_2017_us_county_LOWER_48.shp), which were updated with a column for each mosquito species
Mosquito_Ranges.shp
Presence/absence only was determined.
Note that range maps are out of date, as they are based on Darsie and Ward 1981 and 1989 (from 2001), and are relatively coarse.
No maps of Cx. restuans, although this species is known to be important in WNV transmission
}
\keyword{datasets}
