% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnv_data_docs.R
\docType{data}
\name{us.quarterly}
\alias{us.quarterly}
\title{US Quarterly GRIDMET data}
\format{An object of class \code{data.frame} with 68244 rows and 11 columns.}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/IDAHO_EPSCOR_GRIDMET}
}
\usage{
us.quarterly
}
\description{
Data from the GRIDMET project, downloaded from Google Earth Engine daily by county using the
GRIDMET Viewer and Downloader Version 1.1 tool in the ArboMAP package (www.github.com/ecograph/ArboMAP).
Converted to quarterly data using the RF1 package tool convert.env.data tool designed to take data
from the ArboMAP daily format and put it into the RF1 input format.
convert.env.data is not currently an exported function, but it can be accessed with rf1:::convert.env.data
Currently does not include the anomaly data
}
\details{
Note that the saved R object name is 'us.quarterly' and not 'us_quarter1_gridmet', the .rda file name.
}
\keyword{datasets}
